Add map page
We wrote a Python script to generate `map.html` using
`map.html.template`. The html page has some Javascript to load a big
Google Map, geocode a few sample church addresses, and plot them on the
map. We don't have all the church addresses typed in yet. We started a
`churches.csv` file but the python code doesn't use it yet.

We also have a `families.csv` file that we'll use to find drive time for
each family to each church site. We don't want to publish people's
addresses, so we check in `families.csv.cpt` instead. You can use
`Makefile` targets to get the plaintext, as long as you have the correct
secret key in your `.env` file.