Get driving times for all church/family combinations
Google has the perfect API for this: you give a list of origins and a
list of destinations, and it gives the driving distance & duration for
each element in the matrix. Using it is a little tricky because of
limits of the max dimensions and max total elements. We try to minimize
API calls by finding an optimal group size.

We save the results to a new `trips` table that joins families and
churches.

There are a few details that still need work. See the README, but for
example Google is giving us seconds and we are storing them in a column
named `minutes`!