Show stats for different times & days of week
Added a dropdown for Sunday morning, Saturday evening, and Friday
evening. If you click it then we change the stats in the table. Popups
show stats for that time too.

We had to fix some bugs about how we were getting driving times from
Google. We have to pass `departure_time` not `arrival_time` because
`arrival_time` is ignored for everything but `transit` (i.e. bus, train,
etc.). We still need to update some variables and the database column to
say `departure_time` not `arrival_time`. Also we should probably move
the times we're asking about by about thirty seconds, so they correctly
capture that you'll be leaving during rush hour.

Also if Google is giving you traffic-based drive times, their result
includes both a `duration` value and a `duration_in_traffic` value. So
we should use the `duration_in_traffic` one.

But we had *one* element in the driving time grid where that attribute
was missing. (We don't know why it was missing. I guess that is only one
family/church combo, so not even all results for a family or all results
for a church.) In that case, just fall back to `duration`. Perhaps we
should be printing a warning, but the output is already pretty noisy.

We fixed another bug where calling `(t / 60).toFixed(0)` was giving "-0"
for small negative `t`. Calling `Math.round(t / 60).toString()` doesn't
do that.